[frontend] Split EntityStixCoreRelationships
